@@ -191,7 +191,7 @@ public class ActivityFournisseurs extends AppCompatActivity implements RecyclerA
                         String querry_has_bon1 = "SELECT CODE_FRS FROM ACHAT1 WHERE IS_EXPORTED = 0 AND CODE_FRS = '" + fournisseurs.get(position).code_frs + "'";
                         String querry_has_bon1_temp = "SELECT CODE_FRS FROM ACHAT1_TEMP WHERE IS_EXPORTED = 0 AND CODE_FRS = '" + fournisseurs.get(position).code_frs + "'";
 
-                        if(controller.check_if_fournisseur_has_bon(querry_has_bon1) || controller.check_if_fournisseur_has_bon(querry_has_bon1_temp)){
+                        if(controller.check_if_has_bon(querry_has_bon1) || controller.check_if_has_bon(querry_has_bon1_temp)){
                             // you can't delete this fournisseur
                             new SweetAlertDialog(ActivityFournisseurs.this, SweetAlertDialog.WARNING_TYPE)
                                     .setTitleText("Attention. !")
